+++ services/ui-src/src/views/Qualifiers/Adult/index.tsx
@@ -1,7 +1,8 @@
 import { useEffect } from "react";
 import * as CUI from "@chakra-ui/react";
 import * as QMR from "components";
-import * as Q from "./questions";
+import { DeliverySystems } from "./deliverySystems";
+import * as Common from "../Common";
 import { useForm, FormProvider } from "react-hook-form";
 import { ACSQualifierForm } from "./types";
 import { useParams, useNavigate } from "react-router-dom";
@@ -51,7 +52,7 @@ export const ACSQualifiers = () => {
           userGenerated: false,
         },
       ],
-      CoreSetMeasuresAuditedOrValidatedDetails: [Q.initialAuditValues],
+      CoreSetMeasuresAuditedOrValidatedDetails: [Common.initialAuditValues],
     },
   });
 
@@ -106,10 +107,10 @@ export const ACSQualifiers = () => {
               <QMR.SupportLinks />
             </CUI.Box>
             <CUI.OrderedList>
-              <Q.DeliverySystems />
-              <Q.Audit />
-              <Q.ExternalContractor />
-              <Q.CompleteCoreSets />
+              <DeliverySystems />
+              <Common.Audit type="AD" />
+              <Common.ExternalContractor />
+              <Common.CompleteCoreSets type="AD" />
             </CUI.OrderedList>
           </CUI.Box>
         </form>
