--- services/ui-src/src/views/Qualifiers/Adult/questions/audit.tsx
@@ -1,132 +1,0 @@
-import { useMemo } from "react";
-import * as CUI from "@chakra-ui/react";
-import * as QMR from "components";
-import * as Q from "./";
-import { ICheckbox } from "components/MultiSelect";
-import { useFieldArray } from "react-hook-form";
-import { HiX } from "react-icons/hi";
-import { measuresList } from "measures/measuresList";
-import { useParams } from "react-router-dom";
-
-export const initialAuditValues = {
-  MeasuresAuditedOrValidated: [],
-  WhoConductedAuditOrValidation: "",
-};
-
-export const CloseButton = ({ onClick }: { onClick: () => void }) => (
-  <CUI.IconButton
-    fontSize="1.5em"
-    variant="ghost"
-    icon={<HiX />}
-    aria-label="Remove Audit Item"
-    onClick={onClick}
-  />
-);
-
-export const Audit = () => {
-  const { year } = useParams();
-
-  const { fields, append, remove } = useFieldArray({
-    name: "CoreSetMeasuresAuditedOrValidatedDetails",
-  });
-
-  const adultMeasures = measuresList[year as string]
-    .filter((item) => {
-      return item.type === "AD";
-    })
-    .map((obj) => {
-      return {
-        label: obj.measureId + " - " + obj.name,
-        value: obj.measureId,
-        isVisible: true,
-      };
-    });
-
-  const multiSelectList = useMemo<ICheckbox[]>(
-    () => adultMeasures,
-    [adultMeasures]
-  );
-
-  return (
-    <CUI.ListItem>
-      <Q.QualifierHeader
-        header="Audit or Validation of Measures"
-        description="Were any of the Core Set meaures audited or validated?"
-      />
-      <CUI.Spacer />
-      <CUI.Stack>
-        <CUI.Box pt="4">
-          <QMR.RadioButton
-            formLabelProps={{ fontWeight: "600" }}
-            name="CoreSetMeasuresAuditedOrValidated"
-            options={[
-              {
-                displayValue:
-                  "Yes, some of the Core Set measures have been audited or validated",
-                value:
-                  "Yes, some of the Core Set measures have been audited or validated",
-                children: [
-                  <CUI.Stack mb="5" spacing="6">
-                    {fields?.map((field, index: number) => {
-                      return (
-                        <CUI.Box
-                          border="1px"
-                          borderColor="gray.200"
-                          borderRadius="md"
-                          key={field.id}
-                        >
-                          <CUI.Flex>
-                            <QMR.TextInput
-                              rules={{ required: true }}
-                              formLabelProps={{ fontWeight: "400" }}
-                              label="Who conducted the audit or validation?"
-                              name={`CoreSetMeasuresAuditedOrValidatedDetails.${index}.WhoConductedAuditOrValidation`}
-                              formControlProps={{
-                                p: "5",
-                                pb: "0",
-                              }}
-                            />
-                            <CUI.Spacer />
-                            {index !== 0 && (
-                              <CloseButton onClick={() => remove(index)} />
-                            )}
-                          </CUI.Flex>
-                          <CUI.Box p="5">
-                            <CUI.Text mb="4">
-                              Which measures did they audit or validate?
-                            </CUI.Text>
-
-                            <QMR.MultiSelect
-                              isRequired
-                              multiSelectList={multiSelectList}
-                              name={`CoreSetMeasuresAuditedOrValidatedDetails.${index}.MeasuresAuditedOrValidated`}
-                            />
-                          </CUI.Box>
-                        </CUI.Box>
-                      );
-                    })}
-                  </CUI.Stack>,
-                  <QMR.ContainedButton
-                    buttonText={"+ Add Another"}
-                    buttonProps={{
-                      variant: "outline",
-                      colorScheme: "blue",
-                      textTransform: "capitalize",
-                    }}
-                    onClick={() => append(initialAuditValues)}
-                  />,
-                ],
-              },
-              {
-                displayValue:
-                  "No, none of the Core Set measures have been audited or validated",
-                value:
-                  "No, none of the Core Set measures have been audited or validated",
-              },
-            ]}
-          />
-        </CUI.Box>
-      </CUI.Stack>
-    </CUI.ListItem>
-  );
-};
