+++ services/ui-src/src/views/Qualifiers/Child/deliverySystems.tsx
@@ -1,10 +1,10 @@
 import { useEffect, useState } from "react";
 import * as QMR from "components";
 import * as CUI from "@chakra-ui/react";
-import * as Q from "./";
+import * as Common from "../Common";
 import { useController, useFormContext } from "react-hook-form";
 import { useCustomRegister } from "hooks/useCustomRegister";
-import { DeliverySystem, ACSQualifierForm } from "../types";
+import { DeliverySystem, CCSQualifierForm } from "./types";
 import { BsPercent } from "react-icons/bs";
 import { percentageAllowOneDecimalMax } from "utils/numberInputMasks";
 
@@ -24,7 +24,7 @@ const DefaultDeliverySystemTableRecord = ({
     <CUI.Td>
       <QMR.NumberInput
         displayPercent
-        name={`${record}.TwentyOneToSixtyFour`}
+        name={`${record}.UnderTwentyOneMedicaid`}
         numberInputProps={{ textAlign: "right" }}
         mask={percentageAllowOneDecimalMax}
       />
@@ -32,7 +32,7 @@ const DefaultDeliverySystemTableRecord = ({
     <CUI.Td>
       <QMR.NumberInput
         displayPercent
-        name={`${record}.GreaterThanSixtyFour`}
+        name={`${record}.UnderTwentyOneCHIP`}
         numberInputProps={{ textAlign: "right" }}
         mask={percentageAllowOneDecimalMax}
       />
@@ -47,7 +47,7 @@ interface CustomDeliverySystemTableRecordProps {
 const CustomDeliverySystemTableRecord = ({
   record,
 }: CustomDeliverySystemTableRecordProps) => {
-  const register = useCustomRegister<ACSQualifierForm>();
+  const register = useCustomRegister<CCSQualifierForm>();
   return (
     <CUI.Tr verticalAlign="top">
       <CUI.Td px="none">
@@ -62,7 +62,7 @@ const CustomDeliverySystemTableRecord = ({
       <CUI.Td>
         <QMR.NumberInput
           displayPercent
-          name={`${record}.TwentyOneToSixtyFour`}
+          name={`${record}.UnderTwentyOneMedicaid`}
           numberInputProps={{ textAlign: "right" }}
           mask={percentageAllowOneDecimalMax}
         />
@@ -70,7 +70,7 @@ const CustomDeliverySystemTableRecord = ({
       <CUI.Td>
         <QMR.NumberInput
           displayPercent
-          name={`${record}.GreaterThanSixtyFour`}
+          name={`${record}.UnderTwentyOneCHIP`}
           numberInputProps={{ textAlign: "right" }}
           mask={percentageAllowOneDecimalMax}
         />
@@ -87,19 +87,18 @@ export const DeliverySystems = () => {
   });
 
   const [deliverySystems, setDeliverySystems] = useState(field.value);
-
   const values = watch("PercentageEnrolledInEachDeliverySystem");
 
-  const total21To64Percent = values.reduce(
+  const under21PercentMedicaid = values.reduce(
     (acc: number, curr: DeliverySystem) => {
-      return acc + parseFloat(curr.TwentyOneToSixtyFour || "0");
+      return acc + parseFloat(curr.UnderTwentyOneMedicaid || "0");
     },
     0
   );
 
-  const total65AndOlderPercent = values.reduce(
+  const underTwentyOneCHIP = values.reduce(
     (acc: number, curr: DeliverySystem) => {
-      return acc + parseFloat(curr.GreaterThanSixtyFour || "0");
+      return acc + parseFloat(curr.UnderTwentyOneCHIP || "0");
     },
     0
   );
@@ -123,20 +122,22 @@ export const DeliverySystems = () => {
 
   return (
     <CUI.ListItem>
-      <Q.QualifierHeader
+      <Common.QualifierHeader
         header="Delivery System"
         description="As of September 30, 2021 what percentage of your Medicaid/CHIP
-          enrollees (above age 21) were enrolled in each delivery system?"
+          enrollees (under age 21) were enrolled in each delivery system?"
       />
       <CUI.Table variant="simple" mt="4" size="md">
         <CUI.Thead>
           <CUI.Tr>
             <CUI.Th></CUI.Th>
             <CUI.Th textAlign="center" fontSize="md">
-              <CUI.Text>Ages 21 to 64</CUI.Text>
+              <CUI.Text>Medicaid</CUI.Text>
+              <CUI.Text fontSize="sm">Under Age 21</CUI.Text>
             </CUI.Th>
             <CUI.Th textAlign="center" fontSize="md">
-              <CUI.Text>Age 65 and older</CUI.Text>
+              <CUI.Text>CHIP</CUI.Text>
+              <CUI.Text fontSize="sm">Under Age 21</CUI.Text>
             </CUI.Th>
           </CUI.Tr>
         </CUI.Thead>
@@ -177,7 +178,7 @@ export const DeliverySystems = () => {
               <CUI.InputGroup>
                 <CUI.Input
                   isReadOnly
-                  value={total21To64Percent}
+                  value={under21PercentMedicaid}
                   border="none"
                   textAlign="right"
                   fontWeight="bold"
@@ -194,7 +195,7 @@ export const DeliverySystems = () => {
               <CUI.InputGroup>
                 <CUI.Input
                   isReadOnly
-                  value={total65AndOlderPercent}
+                  value={underTwentyOneCHIP}
                   border="none"
                   textAlign="right"
                   fontWeight="bold"
