--- services/ui-src/src/views/Qualifiers/ChildCHIP/questions/index.tsx
@@ -1,5 +1,0 @@
-export * from "./deliverySystems";
-export * from "./audit";
-export * from "./externalContractor";
-export * from "./complete";
-export * from "./qualifierHeader";
